beggining phase added

+++ Symulacja Cyfrowa/radar_end_of_service_event.cpp
@@ -8,7 +8,7 @@ void RadarEndOfServiceEvent::Execute() {
 	// releasing users in all radar specified channels
 	network_->BandwidthClearRadar();
 	// generating random time of next event in range of 1-5 ms
-	int event_t = radar_time_generator_.RangeRand(1000, 5000) + network_->clock_;
+	int event_t = static_cast<int>(radar_time_generator_.RangeRand(1000, 5000)) + network_->clock_;
 	Event* activation_event = new RadarActivationEvent(event_t, network_, event_list_);
 	event_list_->insert(activation_event);
 	network_->BandwidthPrint();
